backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pymysql==1.1.0
cryptography==42.0.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-dateutil==2.8.2
